pyproject.toml
[tool.poetry]
name = "stocks-exam"
version = "0.1.0"
description = ""
authors = ["Aleksandr <amvorobei@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
pydantic = "^1.9.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
